fix(webapp): scope the jsdom listener shim to the tests that need it
The InvestModal focus-trap test times out on the CI runner and nowhere
else. Its section runs first in the job, so nothing this branch adds is
running before it, and the only change this branch made that reaches it
was patching addEventListener inside the shared setup-dom.

I could not reproduce the hang after matching the runner on OS, bun
version, CPU count, SKIP_ENV_VALIDATION, the exact
`bun test --coverage --coverage-threshold=60` command, and running
`next build` first in the same directory: the test finishes in about 30ms
every time. Rather than keep guessing at a shared file every test in the
workspace loads, the shim moves into its own module that only this
branch's render helper imports.

setup-dom is now byte-identical to develop. Any test that does not render
a component needing the shim gets exactly the DOM behaviour the suite has
been green with.
